requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
alabaster==0.7.12
alembic==1.4.2
appdirs==1.4.4
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1598374659300/work
autopep8==1.5.3
azure-core==1.8.1
azure-storage-blob==12.3.2
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.0
bleach==3.1.5
blinker==1.4
bokeh==2.0.1
boto3==1.14.14
botocore==1.17.63
Bottleneck==1.3.2
branca==0.4.1
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
certipy==0.1.3
cffi @ file:///tmp/build/80754af9/cffi_1598370813909/work
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1588164361239/work
colorama==0.4.3
conda==4.8.4
conda-package-handling==1.7.0
coverage==5.3
coveralls==2.1.2
cryptography @ file:///tmp/build/80754af9/cryptography_1598371423123/work
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.15.0
datascience==0.15.6
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.1
distributed @ file:///tmp/build/80754af9/distributed_1594350110849/work
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
fastcache==1.1.0
filelock==3.0.12
findspark==1.3.0
flake8==3.8.3
folium==0.11.0
fsspec @ file:///tmp/build/80754af9/fsspec_1597944003862/work
future==0.18.2
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.8
gmpy2==2.0.8
gofer-grader==1.1.0
google-auth==1.21.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
HeapDict==1.0.1
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imagesize==1.2.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1593446408836/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1595446871027/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipympl==0.5.6
ipyparallel @ file:///tmp/build/80754af9/ipyparallel_1593440601605/work
ipython @ file:///tmp/build/80754af9/ipython_1598371569012/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
json5==0.9.5
jsonschema @ file:///tmp/build/80754af9/jsonschema_1594363551272/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-conda @ file:///home/conda/feedstock_root/build_artifacts/jupyter_conda_1596994172224/work
jupyter-core==4.6.3
jupyter-lsp==0.8.0
jupyter-server==0.1.1
jupyter-server-proxy==1.5.0
jupyter-telemetry==0.0.5
jupyterhub==1.1.0
jupyterlab==2.1.5
jupyterlab-git==0.21.1
jupyterlab-pygments==0.1.1
jupyterlab-s3-browser==0.6.1
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
Keras==2.1.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
locket==0.2.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1594371495811/work
matplotlib==3.2.1
mccabe==0.6.1
metakernel==0.24.3
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mock==4.0.2
more-itertools==8.5.0
mpmath==1.1.0
msgpack==1.0.0
msrest==0.6.19
multidict==4.7.6
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1598624800312/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1594376811065/work
nbdime==2.0.0
nbformat==5.0.7
nbgitpuller==0.8.0
nbpresent==3.0.2
nbstripout==0.3.9
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
nltk==3.5
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.19.0
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
pamela==1.0.0
pandas==1.0.5
pandocfilters==1.4.2
parso @ file:///tmp/build/80754af9/parso_1596826841367/work
partd==1.1.0
patsy==0.5.1
pexpect @ file:///tmp/build/80754af9/pexpect_1594383317248/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1594384075987/work
Pillow @ file:///tmp/build/80754af9/pillow_1594307325547/work
pipenv==2020.8.13
plotly==4.8.2
pluggy==0.13.1
portalocker @ file:///tmp/build/80754af9/portalocker_1597348414354/work
prometheus-client==0.8.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1598623940270/work
protobuf==3.11.4
psutil @ file:///tmp/build/80754af9/psutil_1598370249042/work
ptyprocess==0.6.0
py==1.9.0
pyarrow==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycodestyle==2.6.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycurl==7.43.0.5
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
pyodbc==4.0.30
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
pyspark==2.4.5
pytest==5.4.1
pytest-cov==2.8.1
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==0.1.11
python-jsonrpc-server==0.3.4
python-language-server==0.34.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.7.14
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.17.0
rsa==4.6
ruamel-yaml==0.15.87
ruamel.yaml.clib @ file:///tmp/build/80754af9/ruamel.yaml.clib_1594403712185/work
s3fs==0.4.2
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
Shapely==1.7.1
simpervisor==0.3
singleton-decorator==1.0.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spylon==0.3.0
spylon-kernel==0.4.1
SQLAlchemy==1.3.16
statsmodels==0.11.1
sympy==1.5.1
tables==3.6.1
tblib @ file:///tmp/build/80754af9/tblib_1597928476713/work
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
torch==1.6.0
toree @ https://dist.apache.org/repos/dist/release/incubator/toree/0.3.0-incubating/toree-pip/toree-0.3.0.tar.gz
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1596810128862/work
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
ujson==1.35
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
vincent==0.4.4
virtualenv==20.0.31
virtualenv-clone==0.5.4
voila==0.1.21
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
yapf==0.30.0
yarl==1.6.0
zict==2.0.0
zipp==3.1.0
